Rework test behaviour w.r.t. local_settings.py

+++ mezzanine/bin/runtests.py
@@ -20,19 +20,25 @@ def main(package="mezzanine"):
     os.environ["DJANGO_SETTINGS_MODULE"] = "project_template.test_settings"
     package_path = path_for_import(package)
     project_path = os.path.join(package_path, "project_template")
+    sys.path.insert(0, package_path)
+
+    tmp_files = []
 
     # Create local_settings.py so it is imported by settings.py
     local_settings_path = os.path.join(project_path, "local_settings.py")
-    shutil.copy(local_settings_path + ".template", local_settings_path)
+    if not os.path.exists(local_settings_path):
+        tmp_files.append(local_settings_path)
+        shutil.copy(local_settings_path + ".template", local_settings_path)
 
     test_settings_path = os.path.join(project_path, "test_settings.py")
-
-    sys.path.insert(0, package_path)
     if not os.path.exists(test_settings_path):
         # Import all our normal settings, and make a few test-specific tweaks.
         # require Mezzanine's accounts app, use the md5 password hasher to
         # speed up tests, use in-memory databases, and define an "other" db.
-        test_settings = """
+        tmp_files.append(test_settings_path)
+        with open(test_settings_path, "w") as f:
+            f.write("""
+
 from project_template import settings
 
 globals().update(i for i in settings.__dict__.items() if i[0].isupper())
@@ -50,16 +56,14 @@ DATABASES = {'default': {'ENGINE': 'django.db.backends.sqlite3'}}
 # needs to be turned on for the contrib.auth tests to pass in Django 1.4.
 PASSWORD_HASHERS = ('django.contrib.auth.hashers.MD5PasswordHasher',
                     'django.contrib.auth.hashers.SHA1PasswordHasher')
-"""
 
-        with open(test_settings_path, "w") as f:
-            f.write(test_settings)
+""")
 
-        def cleanup_test_settings():
-            for fn in [test_settings_path, local_settings_path]:
-                os.remove(fn)
-                os.remove(fn + 'c')
-        atexit.register(cleanup_test_settings)
+    def delete_temporary_files():
+        for filename in tmp_files:
+            os.remove(filename)
+            os.remove(filename + 'c')
+    atexit.register(delete_temporary_files)
 
     if django.VERSION >= (1, 7):
         django.setup()
